+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-wbconfiguration
 0.1.8 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added functionality to view a summary of wbconfiguration.
 
 
 0.1.7 (2012-02-13)